+++ test/analysis_options_test.dart
@@ -55,6 +55,32 @@ formatter:
     });
   });
 
+  test('passthrough linter rules', () {
+    final content = updatePassthroughOptions(
+      original: '''
+linter:
+  rules:
+    avoid_relative_lib_imports: true
+    prefer_relative_imports: false
+    public_member_api_docs: false
+''',
+      custom: '''
+linter:
+  rules:
+    - prefer_relative_imports
+''',
+    );
+    expect(json.decode(content), {
+      'linter': {
+        'rules': {
+          'avoid_relative_lib_imports': true,
+          'prefer_relative_imports': true,
+          'public_member_api_docs': false,
+        },
+      },
+    });
+  });
+
   test('ignore include in original', () {
     final content = updatePassthroughOptions(
       original: 'include: package:lints/other.yaml',