+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.23.1
+
+- Support passthrough of unrelated linter rules in analysis option customization.
+
 ## 0.23.0
 
 **Breaking changes:**